Extract useCourseMeta hook
