+++ assets/js/admin/course-video-sidebar.js
@@ -4,23 +4,11 @@
 import { PluginDocumentSettingPanel } from '@wordpress/edit-post';
 import { __ } from '@wordpress/i18n';
 import { ToggleControl } from '@wordpress/components';
-import { useEntityProp } from '@wordpress/core-data';
 
 /**
- * A hook that provides a value from course meta and a setter for that value.
- *
- * @param {string} metaName The name of the meta.
- *
- * @return {Array} An array containing the value and the setter.
+ * Internal dependencies
  */
-const useCourseMeta = ( metaName ) => {
-	const [ meta, setMeta ] = useEntityProp( 'postType', 'course', 'meta' );
-
-	const value = meta[ metaName ];
-	const setter = ( newValue ) => setMeta( { [ metaName ]: newValue } );
-
-	return [ value, setter ];
-};
+import useCourseMeta from '../../react-hooks/use-course-meta';
 
 /**
  * Video-Based Course Sidebar component.
