+++ assets/react-hooks/use-course-meta.js
@@ -1,0 +1,22 @@
+/**
+ * WordPress dependencies
+ */
+import { useEntityProp } from '@wordpress/core-data';
+
+/**
+ * A hook that provides a value from course meta and a setter for that value.
+ *
+ * @param {string} metaName The name of the meta.
+ *
+ * @return {Array} An array containing the value and the setter.
+ */
+const useCourseMeta = ( metaName ) => {
+	const [ meta, setMeta ] = useEntityProp( 'postType', 'course', 'meta' );
+
+	const value = meta[ metaName ];
+	const setter = ( newValue ) => setMeta( { [ metaName ]: newValue } );
+
+	return [ value, setter ];
+};
+
+export default useCourseMeta;
